Save Game Room

+++ Scripts/AI/Koopa.gd
@@ -68,7 +68,7 @@ func _ai_process(delta: float) -> void:
       owner.kill(AliveObject.DEATH_TYPE.FALL, 0)
 
 func _on_kill_zone_enter(b:Node) -> void:
-  if owner.vars['is shell'] && abs(owner.velocity.x) > 0 && b.is_class('KinematicBody2D') && b != owner:
+  if owner.vars['is shell'] && abs(owner.velocity.x) > 0 && b.is_class('KinematicBody2D') && b != owner && b.has_method('kill'):
     b.kill(AliveObject.DEATH_TYPE.FALL, score_mp)
     #AudioServer.get_bus_effect(1,0).pitch_scale = AliveObject.pitch_md[score_mp]
     #print(AudioServer.get_bus_effect(1,0).pitch_scale)
